Split up HTTP
